ai.hohyub.site/rag/requirements.txt
# LangChain packages (RAG)
langchain-core
langchain-community
langchain-postgres
langchain-openai
langchain-huggingface

# Database
sqlalchemy>=2
psycopg[binary]
asyncpg
pgvector

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv
pydantic

# Embeddings
sentence-transformers

# Local LLM & Embeddings (Llama)
transformers>=4.36.0
accelerate
safetensors
peft
datasets
trl>=0.7.0
bitsandbytes-windows

# Unsloth (Llama fine-tuning) - 주석 처리 (필요시 활성화)
# unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git

# AWS S3 (모델 다운로드용)
boto3>=1.28.0

